Implement a v1 of Create Job Bundle callback
This is a replacement for the existing on_create_job_bundle_callback.
This is not how I want it to stay. Ideally this allows adding/modifying steps and parameters on top of the default DCC code.
It feels unwieldy to make the callback user reimplement everything just to modify or add a step.

Signed-off-by: Alex Hughes <ahughesalex@gmail.com>